Move misplaced requests classes from Application to API #60
