+++ src/API/CompanyName.MyMeetings.API/Modules/Meetings/Meetings/AddMeetingAttendeeRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace CompanyName.MyMeetings.API.Modules.Meetings.Meetings
+{
+    public class AddMeetingAttendeeRequest
+    {
+        public int GuestsNumber { get; set; }
+    }
+}
