+++ src/API/CompanyName.MyMeetings.API/Modules/Meetings/Meetings/RemoveMeetingAttendeeRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace CompanyName.MyMeetings.API.Modules.Meetings
+﻿namespace CompanyName.MyMeetings.API.Modules.Meetings.Meetings
 {
     public class RemoveMeetingAttendeeRequest
     {
