--- src/Modules/Meetings/Application/Meetings/AddMeetingAttendee/AddMeetingAttendeeCommandRequest.cs
@@ -1,7 +1,0 @@
-﻿namespace CompanyName.MyMeetings.Modules.Meetings.Application.Meetings.AddMeetingAttendee
-{
-    public class AddMeetingAttendeeCommandRequest
-    {
-        public int GuestsNumber { get; set; }
-    }
-}
